Styles update all and adds (english) translations
