+++ src/components/cores/updateAll/index.css
@@ -10,20 +10,20 @@
   display: flex;
   gap: 10px;
   flex-direction: column;
-  padding:0;
+  padding: 0;
   flex-grow: 1;
 }
 
 .update-all__no-updates{
-  font-size: 5rem;
-  font-family: GamePocket;
-  width: 100%;
+  font-size: 3rem;
+  font-family: GamePocket, sans-serif;
   height: 100%;
   display: flex;
   align-items: center;
   justify-content: center;
   text-align: center;
   line-height: 1em;
+  margin-inline: auto;
 }
 
 .update-all__list-item {
@@ -31,11 +31,11 @@
   column-gap: 10px;
   row-gap: 5px;
   grid-template-areas:
-    "i n u r p"
-    "i v u r p";
+    "i n v u r p"
+    "i c v u r p";
   border: 1px solid black;
-  grid-template-columns: 185px 1fr 80px 80px 80px;
-  grid-template-rows: auto;
+  background-color: rgb(0 0 0 / 10%);
+  grid-template-columns: 185px 1fr max-content 80px 80px 80px;
   will-change: opacity;
   transition: opacity 0.2s;
 }
@@ -59,14 +59,20 @@
   font-size: 1.5rem;
 }
 
+.update-all__list-item-platform{
+  grid-area: c;
+  opacity: 0.8;
+}
+
 .update-all__list-item-version{
   grid-area: v;
   font-weight: bold;
   display: flex;
   gap: 5px;
+  align-self: center;
 }
 
-.update-all__list-item-update, .update-all__list-item-required-files, .update-all__list-item-platform-images {
+.update-all__list-item-update, .update-all__list-item-required-files, .update-all__list-item-platform-files {
   background-color: rgb(0 0 0 / 50%);
   display: flex;
   justify-content: center;
@@ -88,14 +94,14 @@
   grid-area: r;
 }
 
-.update-all__list-item-platform-images{
+.update-all__list-item-platform-files{
   grid-area: p;
 }
 
 .update-all__step-info{
   display: grid;
   grid-template-columns: 1fr 0.5fr;
-  gap: 10px;
+  column-gap: 10px;
   background-color: var(--light-colour);
   border-radius: 15px;
   grid-template-areas:
@@ -105,6 +111,7 @@
   /* border: 1px solid white; */
   overflow: hidden;
   align-items: stretch;
+  flex-grow: 1;
 
   > * {
     overflow: auto;
@@ -115,6 +122,16 @@
 .update-all__step-info-title {
   padding: 0 20px;
   grid-area: i;
+  display: flex;
+  justify-content: space-between;
+  font-family: GamePocket, sans-serif;
+  font-size: 2rem;
+  line-height: 1em;
+  padding-block: 15px;
+  border-bottom: 1px solid rgb(255 255 255 / 60%);
+  margin: 0;
+  background-color: var(--info-colour);
+  height: min-content;
 }
 
 .update-all__core-progress{
